Fix building with gcc-6

+++ libavogadro/src/extensions/symmetry/libmsym/src/orbital.c
@@ -67,13 +67,13 @@ msym_error_t orbitalFromQuantumNumbers(int n, int l, int m, msym_orbital_t *o){
         }
         case 2 : {
             //o->v = dpolynomial[m+l];
-            char *d = (signbit(m) == 1 ? "-" : "+");
+            char *d = (signbit((float)m) == 1 ? "-" : "+");
             snprintf(o->name, sizeof(o->name), "%dd%d%s",n,abs(m),d);
             break;
         }
         default : {
             char t = 'f' - 3 + l;
-            char *d = (signbit(m) == 1 ? "-" : "+");
+            char *d = (signbit((float)m) == 1 ? "-" : "+");
             snprintf(o->name, sizeof(o->name), "%d%c%d%s",n,t,abs(m),d);
         }
     }
